Replace spaces with %20
Replace spaces with %20, since something in the later code isn't liking separate words. Take the word "top hat", which returns an error through this website but returns a valid definition through the API. I have no way of testing whether what I've done works, but it should. Give it a run, see if it works